+++ CHANGELOG.md
@@ -5,10 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `:Po jump next` and `:Po jump prev` commands to jump to the next or previous untranslated or fuzzy message
+
 ## [0.1.0] - 2024-10-11
 
 ### Added
 
 - Component for [lualine.nvim](https://github.com/nvim-lualine/lualine.nvim) which prints the current translation status of the PO file
 
+[Unreleased]: https://github.com/toku-sa-n/po.nvim/compare/0.1.0...HEAD
 [0.1.0]: https://github.com/toku-sa-n/po.nvim/releases/tag/0.1.0